Update workaround code
After taking a look at
<https://github.com/react-dropzone/react-dropzone/issues/1415#issuecomment-3124165912>,

I realized that I do not need a 'dropzoneRootRef' at all.

I also confirmed that if I drag any file, whether it's an acceptable format or not,
I get a 'isDragActive' state of true.

Therefore, I decided to use this method rather than the ref method.

For end users: there should not be any changes needed to you.

Signed-off-by: seyoon20087 <24852454+seyoon20087@users.noreply.github.com>